# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "user@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/friend/list": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get all friends ",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "friendList",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful show"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/friend/request/list": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Show all friends who request to you ",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "indexRequest",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful show"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/friend/create": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Add friend ",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "storeFriend",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful show"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/friend/confirm": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Confirm friend ",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "confirm",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful show"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/friend/request/remove": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Remove friend request to add friend and delete friend",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "removeFriendRequest",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful show"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
    }
}